+++ test-assets/run-client-tests.sh
@@ -3,7 +3,8 @@ set -euo pipefail
 
 REPO_ROOT=$(git rev-parse --show-toplevel)
 TMPDIR=$(mktemp -d)
-docker=${docker:-docker}
+export docker=${docker:-docker}
+export SERVER_CONTAINER=tacacs-server
 
 if [ ! -v CI ]; then
     # build server image
@@ -30,14 +31,14 @@ test_against_server_image() {
     echo "Testing against image: $image"
 
     echo "Running server container in background"
-    $docker run --rm --detach --publish 5555:5555 --name tacacs-server "$image" >/dev/null
+    $docker run --rm --detach --publish 5555:5555 --name $SERVER_CONTAINER "$image" >/dev/null
 
-    # run all integration tests against server
+    # run integration tests against server
     echo "Running tests..."
     cargo test --package tacacs-plus --test '*' --no-fail-fast
 
     # copy accounting file out of container
-    $docker cp tacacs-server:/tmp/accounting.log $TMPDIR/accounting.log
+    $docker cp $SERVER_CONTAINER:/tmp/accounting.log $TMPDIR/accounting.log
 
     # verify contents of accounting file, printing if invalid
     if ! $REPO_ROOT/test-assets/validate_accounting_file.py $TMPDIR/accounting.log; then
@@ -45,6 +46,9 @@ test_against_server_image() {
         cat $TMPDIR/accounting.log
         return 1
     fi
+
+    # test reconnection by restarting server mid test-run
+    cargo test --package tacacs-plus --test pap_login connection_reestablishment -- --ignored
 }
 
 trap "stop_running_containers" EXIT